requirements.txt
python-telegram-bot==21.6
PyYAML==6.0.2
